+++ src/jsonyx/test/test_run_select_query.py
@@ -29,7 +29,6 @@ class _Slicer:
 _slicer: _Slicer = _Slicer()
 
 
-# TODO(Nice Zombies): test with None
 @pytest.mark.parametrize(("node", "keep"), [
     # List
     (([], slice(0)), True),